Trait rusoto_cognito_sync::CognitoSync

pub trait CognitoSync {
    fn bulk_publish(
        &self, 
        input: BulkPublishRequest
    ) -> RusotoFuture<BulkPublishResponse, BulkPublishError>;

    fn delete_dataset(
        &self, 
        input: DeleteDatasetRequest
    ) -> RusotoFuture<DeleteDatasetResponse, DeleteDatasetError>;

    fn describe_dataset(
        &self, 
        input: DescribeDatasetRequest
    ) -> RusotoFuture<DescribeDatasetResponse, DescribeDatasetError>;

    fn describe_identity_pool_usage(
        &self, 
        input: DescribeIdentityPoolUsageRequest
    ) -> RusotoFuture<DescribeIdentityPoolUsageResponse, DescribeIdentityPoolUsageError>;

    fn describe_identity_usage(
        &self, 
        input: DescribeIdentityUsageRequest
    ) -> RusotoFuture<DescribeIdentityUsageResponse, DescribeIdentityUsageError>;

    fn get_bulk_publish_details(
        &self, 
        input: GetBulkPublishDetailsRequest
    ) -> RusotoFuture<GetBulkPublishDetailsResponse, GetBulkPublishDetailsError>;

    fn get_cognito_events(
        &self, 
        input: GetCognitoEventsRequest
    ) -> RusotoFuture<GetCognitoEventsResponse, GetCognitoEventsError>;

    fn get_identity_pool_configuration(
        &self, 
        input: GetIdentityPoolConfigurationRequest
    ) -> RusotoFuture<GetIdentityPoolConfigurationResponse, GetIdentityPoolConfigurationError>;

    fn list_datasets(
        &self, 
        input: ListDatasetsRequest
    ) -> RusotoFuture<ListDatasetsResponse, ListDatasetsError>;

    fn list_identity_pool_usage(
        &self, 
        input: ListIdentityPoolUsageRequest
    ) -> RusotoFuture<ListIdentityPoolUsageResponse, ListIdentityPoolUsageError>;

    fn list_records(
        &self, 
        input: ListRecordsRequest
    ) -> RusotoFuture<ListRecordsResponse, ListRecordsError>;

    fn register_device(
        &self, 
        input: RegisterDeviceRequest
    ) -> RusotoFuture<RegisterDeviceResponse, RegisterDeviceError>;

    fn set_cognito_events(
        &self, 
        input: SetCognitoEventsRequest
    ) -> RusotoFuture<(), SetCognitoEventsError>;

    fn set_identity_pool_configuration(
        &self, 
        input: SetIdentityPoolConfigurationRequest
    ) -> RusotoFuture<SetIdentityPoolConfigurationResponse, SetIdentityPoolConfigurationError>;

    fn subscribe_to_dataset(
        &self, 
        input: SubscribeToDatasetRequest
    ) -> RusotoFuture<SubscribeToDatasetResponse, SubscribeToDatasetError>;

    fn unsubscribe_from_dataset(
        &self, 
        input: UnsubscribeFromDatasetRequest
    ) -> RusotoFuture<UnsubscribeFromDatasetResponse, UnsubscribeFromDatasetError>;

    fn update_records(
        &self, 
        input: UpdateRecordsRequest
    ) -> RusotoFuture<UpdateRecordsResponse, UpdateRecordsError>;
}

Trait representing the capabilities of the Amazon Cognito Sync API. Amazon Cognito Sync clients implement this trait.

Required methods

fn bulk_publish(
&self,
input: BulkPublishRequest
) -> RusotoFuture<BulkPublishResponse, BulkPublishError>

Initiates a bulk publish of all existing datasets for an Identity Pool to the configured stream. Customers are limited to one successful bulk publish per 24 hours. Bulk publish is an asynchronous request, customers can see the status of the request via the GetBulkPublishDetails operation.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

fn delete_dataset(
&self,
input: DeleteDatasetRequest
) -> RusotoFuture<DeleteDatasetResponse, DeleteDatasetError>

Deletes the specific dataset. The dataset will be deleted permanently, and the action can't be undone. Datasets that this dataset was merged with will no longer report the merge. Any subsequent operation on this dataset will result in a ResourceNotFoundException.

This API can be called with temporary user credentials provided by Cognito Identity or with developer credentials.

fn describe_dataset(
&self,
input: DescribeDatasetRequest
) -> RusotoFuture<DescribeDatasetResponse, DescribeDatasetError>

Gets meta data about a dataset by identity and dataset name. With Amazon Cognito Sync, each identity has access only to its own data. Thus, the credentials used to make this API call need to have access to the identity data.

This API can be called with temporary user credentials provided by Cognito Identity or with developer credentials. You should use Cognito Identity credentials to make this API call.

fn describe_identity_pool_usage(
&self,
input: DescribeIdentityPoolUsageRequest
) -> RusotoFuture<DescribeIdentityPoolUsageResponse, DescribeIdentityPoolUsageError>

Gets usage details (for example, data storage) about a particular identity pool.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

fn describe_identity_usage(
&self,
input: DescribeIdentityUsageRequest
) -> RusotoFuture<DescribeIdentityUsageResponse, DescribeIdentityUsageError>

Gets usage information for an identity, including number of datasets and data usage.

This API can be called with temporary user credentials provided by Cognito Identity or with developer credentials.

fn get_bulk_publish_details(
&self,
input: GetBulkPublishDetailsRequest
) -> RusotoFuture<GetBulkPublishDetailsResponse, GetBulkPublishDetailsError>

Get the status of the last BulkPublish operation for an identity pool.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

fn get_cognito_events(
&self,
input: GetCognitoEventsRequest
) -> RusotoFuture<GetCognitoEventsResponse, GetCognitoEventsError>

Gets the events and the corresponding Lambda functions associated with an identity pool.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

fn get_identity_pool_configuration(
&self,
input: GetIdentityPoolConfigurationRequest
) -> RusotoFuture<GetIdentityPoolConfigurationResponse, GetIdentityPoolConfigurationError>

Gets the configuration settings of an identity pool.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

fn list_datasets(
&self,
input: ListDatasetsRequest
) -> RusotoFuture<ListDatasetsResponse, ListDatasetsError>

Lists datasets for an identity. With Amazon Cognito Sync, each identity has access only to its own data. Thus, the credentials used to make this API call need to have access to the identity data.

ListDatasets can be called with temporary user credentials provided by Cognito Identity or with developer credentials. You should use the Cognito Identity credentials to make this API call.

fn list_identity_pool_usage(
&self,
input: ListIdentityPoolUsageRequest
) -> RusotoFuture<ListIdentityPoolUsageResponse, ListIdentityPoolUsageError>

Gets a list of identity pools registered with Cognito.

ListIdentityPoolUsage can only be called with developer credentials. You cannot make this API call with the temporary user credentials provided by Cognito Identity.

fn list_records(
&self,
input: ListRecordsRequest
) -> RusotoFuture<ListRecordsResponse, ListRecordsError>

Gets paginated records, optionally changed after a particular sync count for a dataset and identity. With Amazon Cognito Sync, each identity has access only to its own data. Thus, the credentials used to make this API call need to have access to the identity data.

ListRecords can be called with temporary user credentials provided by Cognito Identity or with developer credentials. You should use Cognito Identity credentials to make this API call.

fn register_device(
&self,
input: RegisterDeviceRequest
) -> RusotoFuture<RegisterDeviceResponse, RegisterDeviceError>

Registers a device to receive push sync notifications.

This API can only be called with temporary credentials provided by Cognito Identity. You cannot call this API with developer credentials.

fn set_cognito_events(
&self,
input: SetCognitoEventsRequest
) -> RusotoFuture<(), SetCognitoEventsError>

Sets the AWS Lambda function for a given event type for an identity pool. This request only updates the key/value pair specified. Other key/values pairs are not updated. To remove a key value pair, pass a empty value for the particular key.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

fn set_identity_pool_configuration(
&self,
input: SetIdentityPoolConfigurationRequest
) -> RusotoFuture<SetIdentityPoolConfigurationResponse, SetIdentityPoolConfigurationError>

Sets the necessary configuration for push sync.

This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.

fn subscribe_to_dataset(
&self,
input: SubscribeToDatasetRequest
) -> RusotoFuture<SubscribeToDatasetResponse, SubscribeToDatasetError>

Subscribes to receive notifications when a dataset is modified by another device.

This API can only be called with temporary credentials provided by Cognito Identity. You cannot call this API with developer credentials.

fn unsubscribe_from_dataset(
&self,
input: UnsubscribeFromDatasetRequest
) -> RusotoFuture<UnsubscribeFromDatasetResponse, UnsubscribeFromDatasetError>

Unsubscribes from receiving notifications when a dataset is modified by another device.

This API can only be called with temporary credentials provided by Cognito Identity. You cannot call this API with developer credentials.

fn update_records(
&self,
input: UpdateRecordsRequest
) -> RusotoFuture<UpdateRecordsResponse, UpdateRecordsError>

Posts updates to records and adds and deletes records for a dataset and user.

The sync count in the record patch is your last known sync count for that record. The server will reject an UpdateRecords request with a ResourceConflictException if you try to patch a record with a new value but a stale sync count.

For example, if the sync count on the server is 5 for a key called highScore and you try and submit a new highScore with sync count of 4, the request will be rejected. To obtain the current sync count for a record, call ListRecords. On a successful update of the record, the response returns the new sync count for that record. You should present that sync count the next time you try to update that same record. When the record does not exist, specify the sync count as 0.

This API can be called with temporary user credentials provided by Cognito Identity or with developer credentials.

Loading content...