Crate rusoto_cognito_sync

Amazon Cognito Sync
Amazon Cognito Sync provides an AWS service and client library that enable cross-device syncing of application-related user data. High-level client libraries are available for both iOS and Android. You can use these libraries to persist data locally so that it's available even if the device is offline. Developer credentials don't need to be stored on the mobile device to access the service. You can use Amazon Cognito to obtain a normalized user ID and credentials. User data is persisted in a dataset that can store up to 1 MB of key-value pairs, and you can have up to 20 datasets per user identity.

With Amazon Cognito Sync, the data stored for each identity is accessible only to credentials assigned to that identity. In order to use the Cognito Sync service, you need to make API calls using credentials retrieved with Amazon Cognito Identity service.

If you want to use Cognito Sync in an Android or iOS application, you will probably want to make API calls via the AWS Mobile SDK. To learn more, see the Developer Guide for Android and the Developer Guide for iOS.

If you're using the service, you're probably looking for CognitoSyncClient and CognitoSync.

Structs

BulkPublishRequest
The input for the BulkPublish operation.

BulkPublishResponse
The output for the BulkPublish operation.

CognitoStreams
Configuration options for configure Cognito streams.

CognitoSyncClient
A client for the Amazon Cognito Sync API.

Dataset
A collection of data for an identity pool. An identity pool can have multiple datasets. A dataset is per identity and can be general or associated with a particular entity in an application (like a saved game). Datasets are automatically created if they don't exist. Data is synced by dataset, and a dataset can hold up to 1MB of key-value pairs.

DeleteDatasetRequest
A request to delete the specific dataset.

DeleteDatasetResponse
Response to a successful DeleteDataset request.

DescribeDatasetRequest
A request for meta data about a dataset (creation date, number of records, size) by owner and dataset name.

DescribeDatasetResponse
Response to a successful DescribeDataset request.

DescribeIdentityPoolUsageRequest
A request for usage information about the identity pool.

DescribeIdentityPoolUsageResponse
Response to a successful DescribeIdentityPoolUsage request.

DescribeIdentityUsageRequest
A request for information about the usage of an identity pool.

DescribeIdentityUsageResponse
The response to a successful DescribeIdentityUsage request.

GetBulkPublishDetailsRequest
The input for the GetBulkPublishDetails operation.

GetBulkPublishDetailsResponse
The output for the GetBulkPublishDetails operation.

GetCognitoEventsRequest
A request for a list of the configured Cognito Events

GetCognitoEventsResponse
The response from the GetCognitoEvents request

GetIdentityPoolConfigurationRequest
The input for the GetIdentityPoolConfiguration operation.

GetIdentityPoolConfigurationResponse
The output for the GetIdentityPoolConfiguration operation.

IdentityPoolUsage
Usage information for the identity pool.

IdentityUsage
Usage information for the identity.

ListDatasetsRequest
Request for a list of datasets for an identity.

ListDatasetsResponse
Returned for a successful ListDatasets request.

ListIdentityPoolUsageRequest
A request for usage information on an identity pool.

ListIdentityPoolUsageResponse
Returned for a successful ListIdentityPoolUsage request.

ListRecordsRequest
A request for a list of records.

ListRecordsResponse
Returned for a successful ListRecordsRequest.

PushSync
Configuration options to be applied to the identity pool.

Record
The basic data structure of a dataset.

RecordPatch
An update operation for a record.

RegisterDeviceRequest
A request to RegisterDevice.

RegisterDeviceResponse
Response to a RegisterDevice request.

SetCognitoEventsRequest
A request to configure Cognito Events

SetIdentityPoolConfigurationRequest
The input for the SetIdentityPoolConfiguration operation.

SetIdentityPoolConfigurationResponse
The output for the SetIdentityPoolConfiguration operation

SubscribeToDatasetRequest
A request to SubscribeToDatasetRequest.

SubscribeToDatasetResponse
Response to a SubscribeToDataset request.

UnsubscribeFromDatasetRequest
A request to UnsubscribeFromDataset.

UnsubscribeFromDatasetResponse
Response to an UnsubscribeFromDataset request.

UpdateRecordsRequest
A request to post updates to records or add and delete records for a dataset and user.

UpdateRecordsResponse
Returned for a successful UpdateRecordsRequest.

Enums

BulkPublishError
Errors returned by BulkPublish

DeleteDatasetError
Errors returned by DeleteDataset

DescribeDatasetError
Errors returned by DescribeDataset

DescribeIdentityPoolUsageError
Errors returned by DescribeIdentityPoolUsage

DescribeIdentityUsageError
Errors returned by DescribeIdentityUsage

GetBulkPublishDetailsError
Errors returned by GetBulkPublishDetails

GetCognitoEventsError
Errors returned by GetCognitoEvents

GetIdentityPoolConfigurationError
Errors returned by GetIdentityPoolConfiguration

ListDatasetsError
Errors returned by ListDatasets

ListIdentityPoolUsageError
Errors returned by ListIdentityPoolUsage

ListRecordsError
Errors returned by ListRecords

RegisterDeviceError
Errors returned by RegisterDevice

SetCognitoEventsError
Errors returned by SetCognitoEvents

SetIdentityPoolConfigurationError
Errors returned by SetIdentityPoolConfiguration

SubscribeToDatasetError
Errors returned by SubscribeToDataset

UnsubscribeFromDatasetError
Errors returned by UnsubscribeFromDataset

UpdateRecordsError
Errors returned by UpdateRecords

Traits

CognitoSync
Trait representing the capabilities of the Amazon Cognito Sync API. Amazon Cognito Sync clients implement this trait.