Struct rusoto_cognito_sync::CognitoSyncClient

pub struct CognitoSyncClient { /* fields omitted */ }

A client for the Amazon Cognito Sync API.

Methods

impl CognitoSyncClient

pub fn new(region: Region) -> CognitoSyncClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CognitoSyncClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,